+++ changelog.md
@@ -29,6 +29,7 @@ Codion Change Log
 - AbstractComponentValueBuilder.createComponentValue() renamed createValue().
 - ComponentValueBuilder, a few missing listener and consumer methods added.
 - AbstractComponentValueBuilder, listener management now consistent with DefaultValue.DefaultBuilder.
+- FilterTable.FilterTableHeader removed, now uses default JTableHeader toolTip functionality, COLUMN_TOOL_TIPS configuration value added.
 ### is.codion.framework.model
 - AttributeConditionModel and ColumnConditionModel added.
 - DefaultEntityConditionModel, temporal equal precision adjustment now half-open interval based.